Add the JobDecorator abstraction using Draper
