+++ Gemfile
@@ -31,6 +31,7 @@ gem 'commander-openflighthpc', '~> 2.1'
 gem 'activemodel'
 gem 'flight_configuration', github: 'openflighthpc/flight_configuration', tag: '0.5.0'
 gem 'dotenv'
+gem 'draper'
 gem 'json_schemer'
 gem 'output_mode'
 gem 'pastel'
